+++ src/config/pragmas/sort/http.js
@@ -11,18 +11,15 @@ let { httpMethods } = require('../../../lib')
  * - Finally, ensure root captures rank below a root literal
  */
 module.exports = function sortHTTP (http) {
+  let param = /\/:/
+
   // Construct the tree from HTTP methods
   let tree = {}
   http.forEach(({ method, path }) => {
     if (!tree[method]) tree[method] = []
     let parts = path.split('/').filter(Boolean)
     let depth = parts.length
-    let item = { depth, path }
-    let param = /\/:/
-    if (path.match(param)) {
-      item.hasParam = true
-      item.paramIndex = path.match(param).index // If multiple, we want the earliest
-    }
+    let item = { depth, path, parts, hasParam: param.test(path) }
     if (parts.length) {
       if (parts[depth - 1] === '*') item.trailingCapture = 'catchall'
       if (parts[depth - 1].startsWith(':')) item.trailingCapture = 'param'
@@ -40,18 +37,28 @@ module.exports = function sortHTTP (http) {
       .sort((a, b) => b.depth - a.depth)
       // Sort within a given depth
       .sort((a, b) => {
-        // Handle root (depth: 0)
+        // Depth is already correct
         if (a.depth - b.depth < 0) return
-        if (a.hasParam && b.hasParam) {
-          // Sort at the earliest param
-          if (a.paramIndex < b.paramIndex) return 1
-          if (a.paramIndex > b.paramIndex) return -1
-          // Then sort alphabetically
-          if (a.path < b.path) return -1
-          if (a.path > b.path) return 1
+        // jic we aren't in the same depth (this prob won't be run)
+        if (a.depth !== b.depth) return b.depth - a.depth
+        // Sort static parts vs  params by position in path
+        if (a.hasParam || b.hasParam) {
+          for (let i = 0; i < a.depth; i++) {
+            let aPart = a.parts[i]
+            let bPart = b.parts[i]
+            let aPartIsParam = aPart.startsWith(':')
+            let bPartIsParam = bPart.startsWith(':')
+            // If position is both params, move onto the next position
+            if (aPartIsParam && bPartIsParam) continue
+            // Static parts get priority over params
+            if (aPartIsParam) return 1
+            if (bPartIsParam) return -1
+            // Sort alphabetically
+            if (aPart < bPart) return -1
+            if (aPart > bPart) return 1
+          }
         }
-        if (a.hasParam) return 1
-        if (b.hasParam) return -1
+        // Final alphabetical sort
         if (a.path < b.path) return -1
         if (a.path > b.path) return 1
       })