+++ src/env/index.js
@@ -11,7 +11,8 @@ module.exports = function env (params, inventory, callback) {
   /* istanbul ignore next */
   try {
     // eslint-disable-next-line
-    require('aws-sdk/lib/maintenance_mode_message').suppress = true
+    try { require('aws-sdk/lib/maintenance_mode_message').suppress = true }
+    catch { /* Noop */ }
     // eslint-disable-next-line
     var aws = require('aws-sdk')
   }