+++ test/unit/src/config/project/plugins/runtimes-test.js
@@ -114,23 +114,56 @@ test('Transpiled runtime setters', t => {
   t.equal(plugins.build, build, 'Returned explicit build property')
 })
 
-test('Runtime setter validation (transpiled)', t => {
-  t.plan(2)
-  let inventory, plugin, runtime
-  let type = 'transpiled'
+test('Compiled runtime setters', t => {
+  t.plan(14)
+  let inventory, plugin, plugins, runtime
+  let type = 'compiled'
+  let build = '.build' // Just make sure it's different from the default 'build'
+  let defaultBaseRuntime = 'provided.al2'
+  let baseRuntime = 'provided'
   let errors = []
 
-  // Invalid baseRuntime
-  runtime = { name, type, baseRuntime: 'nodejs10.x' }
+  // Compiled with default build dir
+  runtime = { name, type }
   plugin = () => runtime
   inventory = newInv([ plugin ])
-  setRuntimesPlugins({ inventory, errors }, emptyProj)
-  t.equal(errors.length, 1, 'Returned an error')
-  t.match(errors[0], /must include a valid baseRuntime property/, 'Returned correct baseRuntime error')
+  plugins = setRuntimesPlugins({ inventory, errors }, emptyProj)
+  t.notOk(errors.length, 'Did not return errors')
+  t.equal(plugins.build, 'build', 'Returned default build property')
+  t.equal(plugins.runtimes.runtimes.length, 1, 'Returned single runtime')
+  t.equal(plugins.runtimes.runtimes[0], name, 'Returned correct runtime name')
+  t.equal(plugins.runtimes[name].name, name, 'Returned populated custom runtime object')
+  t.deepEqual(plugins.runtimes[name], runtime, 'Returned populated custom runtime object')
+  t.equal(plugins.runtimes[name].baseRuntime, defaultBaseRuntime, 'Returned default baseRuntime')
+
+  // Compiled with specified baseRuntime
+  runtime = { name, type, baseRuntime }
+  plugin = () => runtime
+  inventory = newInv([ plugin ])
+  plugins = setRuntimesPlugins({ inventory, errors }, emptyProj)
+  t.notOk(errors.length, 'Did not return errors')
+  t.equal(plugins.build, 'build', 'Returned default build property')
+  t.equal(plugins.runtimes[name].baseRuntime, baseRuntime, 'Returned correct baseRuntime')
+
+  // Compiled with default build dir via truthy build property
+  runtime = { name, type, build: true }
+  plugin = () => runtime
+  inventory = newInv([ plugin ])
+  plugins = setRuntimesPlugins({ inventory, errors }, emptyProj)
+  t.notOk(errors.length, 'Did not return errors')
+  t.equal(plugins.build, 'build', 'Returned default build property')
+
+  // Compiled with default build dir via explicit build property
+  runtime = { name, type, build }
+  plugin = () => runtime
+  inventory = newInv([ plugin ])
+  plugins = setRuntimesPlugins({ inventory, errors }, emptyProj)
+  t.notOk(errors.length, 'Did not return errors')
+  t.equal(plugins.build, build, 'Returned explicit build property')
 })
 
 test('Runtime setter validation', t => {
-  t.plan(14)
+  t.plan(16)
   let errors, inventory, plugin, runtime
   let type = 'transpiled'
   let baseRuntime = 'nodejs14.x'
@@ -196,6 +229,15 @@ test('Runtime setter validation', t => {
   setRuntimesPlugins({ inventory, errors }, emptyProj)
   t.equal(errors.length, 1, 'Returned an error')
   t.match(errors[0], /cannot set a build directory, as it is already configured/, 'Returned correct build error')
+
+  // Invalid baseRuntime
+  runtime = { name, type, baseRuntime: 'nodejs10.x' }
+  plugin = () => runtime
+  inventory = newInv([ plugin ])
+  errors = []
+  setRuntimesPlugins({ inventory, errors }, emptyProj)
+  t.equal(errors.length, 1, 'Returned an error')
+  t.match(errors[0], /must include a valid baseRuntime property/, 'Returned correct baseRuntime error')
 })
 
 test('Runtime setter errors', t => {